+++ rest_framework/views/aiohttp/views.py
@@ -5,15 +5,33 @@ Views for aiohttp.
 import asyncio
 
 from aiohttp.hdrs import METH_ALL
+from aiohttp.web import (
+    View as AioHttpClassBaseView,
+    json_response
+)
 
-from rest_framework.views import BaseApiView, GetSerializerAbstractMixin
+from rest_framework.views.base import BaseApiView
+from rest_framework.exceptions import ApiException
 
 
-class AioHTTPApiView(BaseApiView, GetSerializerAbstractMixin):
+class AioHTTPApiView(AioHttpClassBaseView, BaseApiView):
     """
     AioHTTP base API view.
 
     """
+    response_class = json_response
+
+    @property
+    def request_object(self):
+        """
+        Get request object.
+
+        :return: Request object.
+        :rtype:
+
+        """
+        return self.request
+
     @property
     def current_request_method(self):
         """
@@ -39,6 +57,21 @@ class AioHTTPApiView(BaseApiView, GetSerializerAbstractMixin):
         if method is None:
             self._raise_allowed_methods()
 
-        response = await self.dispatch(method)
+        response = yield from self.dispatch(method)
 
         return response
+
+    @asyncio.coroutine
+    def dispatch(self, method):
+        """
+        Code after, before call request handler.
+
+        :param function method: Method handler for call.
+
+        """
+        try:
+            result = yield from method()
+            return result
+        except ApiException as e:
+            # TODO: Not security. e.detail maybe anything
+            return self.response_class({'errors': e.detail}, status=e.status or 400)